Added script to list states with names starting with N from the database hbtn_0e_0_usa
Created a Python script that connects to the database hbtn_0e_0_usa and retrieves a list of all states with names starting with the letter 'N'. The script takes three command-line arguments: the database username, password, and name. The script connects to the database using the provided credentials and executes the SQL query to retrieve the states. The retrieved states are then printed to the console.