Added script to list all states from the database hbtn_0e_0_usa
Created a Python script that connects to the database hbtn_0e_0_usa and retrieves a list of all states. The script prompts the user for the database username, password, and name before connecting to the database and executing the SQL query to retrieve the states. The retrieved states are then printed to the console.