Added script to send POST request with email parameter and display decoded response
Implemented a Python script that sends a POST request to a specified URL with an email address as a parameter and displays the decoded body of the response in UTF-8 format. The script accepts the URL and email address as arguments, utilizes the urllib.parse module to encode the parameter data, and constructs an HTTP POST request using the urllib.request module. It then decodes the response body from UTF-8 bytes and displays it to the console.