Added script to fetch and display content from https://alx-intranet.hbtn.io/status
Implemented a Python script that fetches the content from the URL [https://alx-intranet.hbtn.io/status](https://alx-intranet.hbtn.io/status) and displays the raw content of the response. The script utilizes the requests library to make an HTTP GET request to the specified URL and extracts the response content.