Added script to send POST request with email parameter and display response body
Implemented a Python script that utilizes the requests library to send a POST request to a specified URL with an email address as a parameter and displays the decoded body of the response. The script accepts the URL and email address as arguments, constructs an HTTP POST request with the email parameter, and sends the request to the URL. It then decodes the response body and displays it to the console.