Added script to send DELETE request and display response body
Implemented a Bash script that utilizes the curl command to send a DELETE request to a specified URL. The script accepts the URL as an argument and executes a curl command with the -sX DELETE option to send the DELETE request. It then captures the response body and displays it to the console.