Added script to fetch and display allowed HTTP methods for a URL
Implemented a Bash script that utilizes the curl command to send an OPTIONS request to a specified URL and extract the list of allowed HTTP methods from the Allow header. The script accepts the URL as an argument and executes a curl command with the -sI and -X OPTIONS options to retrieve the response headers and specify the OPTIONS method. It then pipes the output of the curl command to the grep command to filter for the Allow header and extracts the list of methods using the cut command. Finally, it prints the extracted list of allowed HTTP methods to the console.