Added script to send custom header with curl and display response body
Implemented a Bash script that utilizes the curl command to send a customized HTTP request with an additional header (X-School-User-Id: 98) to a specified URL. The script accepts the URL as an argument and executes a curl command with the -sH option to set the custom header. It then captures the response body and displays it to the console.