Added script to send POST request with specified parameters and display response body
Implemented a Bash script that utilizes the curl command to send a POST request to a specified URL with predefined parameters (email=test@gmail.com&subject=I will always be here for PLD). The script accepts the URL as an argument and executes a curl command with the -sX POST and -d options to specify the POST method and data parameters. It then captures the response body and displays it to the console.